==> 3. producing_map.ipynb <==
# SPAH Postcode Mapping - Producing map

import pandas as pd
import folium

### Import data

# define file pathways
cleaned_folder_path = r'C:\Users\scotth07\OneDrive - NHS Scotland\Documents\Networks & Projects\SPAH\Postcode Mapping\cleaned_data'
raw_folder_path = r'C:\Users\scotth07\OneDrive - NHS Scotland\Documents\Networks & Projects\SPAH\Postcode Mapping\raw_data'

nearest_hospitals_pathway = cleaned_folder_path + r'\nearest_hospitals.xlsx' 
hospital_pathway = raw_folder_path + r'\hospital_postcodes.xlsx'

# import data
nearest_hospitals_df = pd.read_excel(nearest_hospitals_pathway, sheet_name='Sheet1') 
hospitals_df = pd.read_excel(hospital_pathway, sheet_name='Sheet1')

### Produce map with marker objects showing postcodes

filtered = nearest_hospitals_df[nearest_hospitals_df['closest_hospital'].isna()]

nearest_hospitals_df = nearest_hospitals_df[nearest_hospitals_df['closest_hospital'].notna()]

# initialise map object
patient_map = folium.Map(
    location=[nearest_hospitals_df['patient_postcode_lat'].mean(), nearest_hospitals_df['patient_postcode_long'].mean()], 
    zoom_start=6
)

# define unique markers for each hospital
marker_styles = {
    'Aberdeen Royal Infirmary': {'color': 'pink', 'icon': 'asterisk'},
    'Dr Gray’s Hospital Elgin': {'color': 'lightgreen', 'icon': 'ok-sign'},
    'Dumfries & Galloway Royal Infirmary': {'color': 'black', 'icon': 'plus-sign'},
    'Edinburgh Royal Infirmary': {'color': 'purple', 'icon': 'info-sign'},
    'Forth Valley Royal Hospital': {'color': 'lightgray', 'icon': 'minus-sign'},
    'Glasgow Royal Infirmary': {'color': 'orange', 'icon': 'bell'},
    'Inverclyde Royal Hospital': {'color': 'gray', 'icon': 'chevron-up'},
    'Ninewells Hospital': {'color': 'beige', 'icon': 'resize-vertical'},
    'Queen Elizabeth University Hospital': {'color': 'pink', 'icon': 'unchecked'},
    'Royal Alexandra Hospital': {'color': 'lightgreen', 'icon': 'question-sign'}
    'St John’s Hospital Livingston': {'color': 'purple', 'icon': 'ok-sign'},
    'University Hospital Ayr': {'color': 'lightgray', 'icon': 'plus-sign'},
    'University Hospital Crosshouse': {'color': 'orange', 'icon': 'minus-sign'},
    'University Hospital Hairmyres': {'color': 'gray', 'icon': 'bell'},
    'University Hospital Monklands': {'color': 'beige', 'icon': 'chevron-up'},
    'University Hospital Wishaw': {'color': 'gray', 'icon': 'resize-vertical'},
    'Victoria Hospital Fife': {'color': 'beige', 'icon': 'unchecked'}
}

# icons {'asterisk', 'ok-sign', 'plus-sign', 'info-sign', 'minus-sign', 'bell', 'chevron-up', 'resize-vertical',
# 'unchecked', 'question-sign'}

# colours {'pink', 'black', 'purple', 'lightgreen', 'lightgray', 'orange', 'gray', 'beige', 
# 'darkgreen', 'cadetblue', 'green', 'lightred', 'blue', 'darkred', 'darkpurple', 'lightblue', 'white', 'darkblue', 'red'}

# add postcode markers to map
for index, location_info in nearest_hospitals_df.iterrows():
    hospital = location_info['closest_hospital']
    style = marker_styles.get(hospital)

    folium.Marker(
        location=[location_info['patient_postcode_lat'], location_info['patient_postcode_long']],
        popup=hospital,
        icon=folium.Icon(color=style['color'], icon=style['icon'])
    ).add_to(patient_map)

patient_map
